scripts/requirements.txt
sentence-transformers~=3.2.1
psycopg2~=2.9.10
python-dotenv~=1.0.1